refactor: 연산 담당 클래스 분리
변경사항
- 기존 class ViewController의 func calculate(expression:) 메소드를 단독 클래스로 분리
